Prototype backup
This update adds no functionally but pastes in the prototype which is runnable on its own but not connected to the the rest of the program yet, it includes a FEA solver for piping systems but not connected to the main software yet.

It also includes a big library for linear algebra. This is a big part of FEM.